requirements.txt
serpapi==1.1.1
streamlit>=1.31.0
python-dotenv>=1.0.0
pymongo[srv]>=4.6.0
groq>=0.4.0 
werkzeug>=2.3.0
certifi>=2024.2.0
resend>=0.6.0
requests>=2.31.0